requirements.txt
flask
flask_caching
flask_mqtt
netaddr
paho-mqtt
pyorbital
